third_party/WebKit/LayoutTests/document-has-focus-expected.txt

CONSOLE MESSAGE: line 8: FOCUS -> document.hasFocus(): true
CONSOLE MESSAGE: line 11: BLUR -> document.hasFocus(): false
CONSOLE MESSAGE: line 8: FOCUS -> document.hasFocus(): true
